django/backend/requirements.txt
asgiref==3.7.2
#backports.zoneinfo==0.2.1
backports.zoneinfo;python_version<"3.9"
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-admin-rangefilter==0.10.0
django-bootstrap5==23.3
django-cors-headers==4.2.0
django-environ==0.10.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
idna==3.4
#mysqlclient==2.2.0
mysqlclient==2.1.1
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing-extensions==4.7.1
#urllib3==2.0.3
urllib3==2.0.2
